Initialise moving-particles in git/Atom
I was working on this project in the editor.p5js.org.

While this approach was great because the results of my coding were
instantaneously visible on the same browser tab, the unfortunate
side-product was that I got lost in working on my project in the web-
editor and, because of this, no longer have access to how the project
evolved in incremenents.

An initial version created four particles at the voronoi generators.
These particles jittered around the screen. They then began to feel
the pull of their home every time they veered out of their zone of
origin. They then began to move by perlin noise. And many other small
incremental changes which I can no longer recall etc..