Particles are aware of whether they are home or not
The particles' .amIHome() method was failing because the if()
condition that checked whether particles' color and the color
of the background that the particle was currently on were the
same was failing.

It was failing because p5.js get() method was returning RGBA with
integer values, and because I had set the voronoi background with
random() which returns floats. I also was not using the A, so: I was
essentially trying to compare RGBA (integer array) with RGB (float array).

Thus, although the particles appeared to be the same color as their
background, they were not. I fixed this by rounding the random(0,255)
that is passed to the voronoi construcor and by adding a fourth value
(of 255 as A).

Despite these modifications, the comparison still didn't work, so I
converted both values .toString() and compared the strings. This works.
And now particles know whether they are home or not.