Particles move towards attractors
With two major improvements:
1) Once attractors emerge, particles are assigned automatically
to attractors. This is implemented as a state of the particles:
this.hasDestiny which is set to either "undefined" or "defined"
(kind of clumsy)

2) Once attractors are removed, particles go back to their own cells.
Note that attractors are not removed from the attractors array,
but they are "inactivated" using a status this.existance which is
set to either "undefined" or "defined".

3) Attractors are now much bigger, and transparent. Particles
hover within the attractors, but also venture a little bit outside,
nevertheless staying close by for the duration of the attraction.
They no longer jitter or go crazy nervous around attractors.