Particle checks its location by comparing cellIDs and other improvements
Particle checks which cell it is located in, and checks whether this is its birth cell. This update in logics paves the way for implementing passport modes. (Old logic was to compare background color to particle color, see next point).

Update color palette with all 12 primary, secondary and tertiary colors. Set alpha to 127 (50% transparent). Particle color blends with background color for effect. Also for this reason, the old logic of checking whether home based on color no longer worked. Anyhow, that logic was in dire need of improvement anyways.

Infobutton, which shows instructions for toggling the "You" particle on mouseover. If clicked, it playsback audio instructions/loop: "if you want something to change you have to speak out"

Speakout-button, which appears sometime after the audio has been triggered. Button implemented such that user needs to hold it down. This button does not currently have functionality implemented, but the idea is to use this to trigger capture of user speech input. (ie.: audio capture is only "active" while user is holding the button down == holding down button as user consent to being listened to by the system)