DEV DEMO: particle movement limited to certain zones
In this dev demo, the movement of particles is limited to certain zones (all cells with an ID number lower than ones own cell). However, if we create an attractor, they can go to the attractor no matter what zone it is located in (next step: do not allow this).

The You particle can be driven using the arrow keys.

For purposes of clarity, this demo version only generates a handful of particles into the world, so that it is easier to observe their behavior.

The only "game" aspect at the moment (user freedom) is to explore the limits of the allowed zones for the You particle. If born into a cell with a low ID number, you soon feel frustrated because your movement within the world is very limited.