Implement going towards attractors
As a bonus, a new idea: once particles make it to their destiny, they eventually pop out of existance along with the destiny ;-) Alas, new particles are always being born. If the user wishes for all the particles in the world to be matched with their destinies, they will have to play forever. Or just leave the borders open.

Add more portuguese destinies.

Slow down particle movement considerably. This allows us to notice when the particles start feeling the force of their destiny (their movement speeds up a lot). Also allows us to notice which particles have a border between themselves and their destiny.