Implement voice interface for all passports in PT and EN
After some iteration, decided to display the text to be spoken more or less in the middle of the screen. The transcript appears in the same position.

Button text flickers between two modes/words, eg: eu só posso/quero ficar aqui, which is a nice idea BUT the flickering is too intense. The problem is: in order to display the text above the zones and particles, we need to draw the text inside draw(), which means that it is being redrawn at each frame.